feat: add tests and adjust tests about payment
