+++ src/contexts/cardPaymentInformationContext/index.tsx
@@ -1,25 +1,8 @@
 import { useCurrentUser } from "contexts/currentUserContext";
-import { useLanguage } from "contexts/languageContext";
-import { coinByLanguage } from "lib/coinByLanguage";
-import {
-  createContext,
-  SetStateAction,
-  useContext,
-  useEffect,
-  useMemo,
-  useState,
-} from "react";
+import { createContext, useContext, useMemo } from "react";
 import { useTranslation } from "react-i18next";
 import { logError } from "services/crashReport";
 import { creditCardPaymentApi } from "@ribon.io/shared/services";
-import { getLocalStorageItem, setLocalStorageItem } from "lib/localStorage";
-import {
-  Cause,
-  Currencies,
-  Languages,
-  NonProfit,
-  Offer,
-} from "@ribon.io/shared/types";
 import { showToast } from "lib/Toast";
 import { useLoadingOverlay } from "contexts/loadingOverlayContext";
 import { useNavigation } from "hooks/useNavigation";
@@ -28,46 +11,12 @@ import { useIntegration, useSources, useUsers } from "@ribon.io/shared/hooks";
 import { normalizedLanguage } from "lib/currentLanguage";
 import { logEvent } from "services/analytics";
 import { useTasksContext } from "contexts/tasksContext";
-import { countryByLanguage } from "lib/countryByLanguage";
 import { useIntegrationContext } from "contexts/integrationContext";
 import { useUtmContext } from "contexts/utmContext";
+import { useCheckoutContext } from "contexts/checkoutContext";
 
 export interface ICardPaymentInformationContext {
-  setCurrentCoin: (value: SetStateAction<Currencies | undefined>) => void;
-  setCountry: (value: SetStateAction<string>) => void;
-  setState: (value: SetStateAction<string>) => void;
-  setCity: (value: SetStateAction<string>) => void;
-  setTaxId: (value: SetStateAction<string>) => void;
-  setEmail: (value: SetStateAction<string | undefined>) => void;
-  setNumber: (value: SetStateAction<string>) => void;
-  setName: (value: SetStateAction<string>) => void;
-  setExpirationDate: (value: SetStateAction<string>) => void;
-  setCvv: (value: SetStateAction<string>) => void;
-  setButtonDisabled: (value: SetStateAction<boolean>) => void;
-  setCryptoGiving: (value: SetStateAction<string>) => void;
-  setOffer: (value: SetStateAction<Offer | undefined>) => void;
-  setFlow: (value: SetStateAction<"cause" | "nonProfit">) => void;
-  buttonDisabled: boolean;
-  currentCoin?: Currencies;
-  country: string;
-  state: string;
-  city: string;
-  taxId: string;
-  email: string | undefined;
-  number: string;
-  name: string;
-  expirationDate: string;
-  cvv: string;
-  cryptoGiving: string;
-  offer: Offer | undefined;
-  flow: "cause" | "nonProfit";
   handleSubmit: () => void;
-  cause: Cause | undefined;
-  setCause: (value: SetStateAction<Cause | undefined>) => void;
-  nonProfit: NonProfit | undefined;
-  setNonProfit: (value: SetStateAction<NonProfit | undefined>) => void;
-  loading: boolean;
-  resetStates: () => void;
 }
 
 export type Props = {
@@ -82,51 +31,8 @@ export const CardPaymentInformationContext =
 export const CURRENT_COIN_KEY = "CURRENT_COIN_KEY";
 
 function CardPaymentInformationProvider({ children }: Props) {
-  const { currentUser } = useCurrentUser();
-  const { currentLang } = useLanguage();
-  const currentContry = countryByLanguage(currentLang);
   const { registerAction } = useTasksContext();
-  const [loading, setLoading] = useState(false);
-  const [currentCoin, setCurrentCoin] = useState<Currencies>();
-  const defaultCoin = async () =>
-    ((await getLocalStorageItem(CURRENT_COIN_KEY)) as Currencies) ||
-    coinByLanguage(currentLang as Languages);
 
-  useEffect(() => {
-    async function setDefaultCoin() {
-      try {
-        setLoading(true);
-        const coin = await defaultCoin();
-        setCurrentCoin(coin);
-      } catch (error) {
-        logError(error);
-      } finally {
-        setLoading(false);
-      }
-    }
-
-    setDefaultCoin();
-  }, []);
-
-  useEffect(() => {
-    if (currentCoin) setLocalStorageItem(CURRENT_COIN_KEY, currentCoin);
-  }, [currentCoin]);
-
-  const [country, setCountry] = useState(currentContry);
-  const [state, setState] = useState("");
-  const [city, setCity] = useState("");
-  const [taxId, setTaxId] = useState("");
-  const [email, setEmail] = useState(currentUser?.email ?? undefined);
-  const [number, setNumber] = useState("");
-  const [name, setName] = useState("");
-  const [expirationDate, setExpirationDate] = useState("");
-  const [cvv, setCvv] = useState("");
-  const [buttonDisabled, setButtonDisabled] = useState(true);
-  const [cryptoGiving, setCryptoGiving] = useState("");
-  const [offer, setOffer] = useState<Offer>();
-  const [cause, setCause] = useState<Cause>();
-  const [nonProfit, setNonProfit] = useState<NonProfit>();
-  const [flow, setFlow] = useState<"nonProfit" | "cause">("nonProfit");
   const { navigateTo } = useNavigation();
 
   const { t } = useTranslation("translation", {
@@ -140,22 +46,23 @@ function CardPaymentInformationProvider({ children }: Props) {
   const { integration } = useIntegration(currentIntegrationId);
   const { createSource } = useSources();
 
-  const resetStates = () => {
-    setCountry("");
-    setEmail(currentUser?.email ?? undefined);
-    setState("");
-    setCity("");
-    setTaxId("");
-    setNumber("");
-    setName("");
-    setExpirationDate("");
-    setCvv("");
-    setButtonDisabled(true);
-  };
-
-  useEffect(() => {
-    if (currentUser) setEmail(currentUser.email);
-  }, [JSON.stringify(currentUser)]);
+  const {
+    email,
+    cvv,
+    name,
+    number,
+    taxId,
+    setButtonDisabled,
+    expirationDate,
+    state,
+    country,
+    city,
+    cause,
+    nonProfit,
+    flow,
+    offer,
+    resetStates,
+  } = useCheckoutContext();
 
   const login = async () => {
     if (!signedIn) {
@@ -236,60 +143,9 @@ function CardPaymentInformationProvider({ children }: Props) {
 
   const cardPaymentInformationObject: ICardPaymentInformationContext = useMemo(
     () => ({
-      currentCoin,
-      setCurrentCoin,
-      country,
-      setCountry,
-      city,
-      setCity,
-      state,
-      setState,
-      taxId,
-      setTaxId,
       handleSubmit,
-      setName,
-      name,
-      setNumber,
-      number,
-      setEmail,
-      email,
-      setExpirationDate,
-      expirationDate,
-      setCvv,
-      cvv,
-      buttonDisabled,
-      setButtonDisabled,
-      setCryptoGiving,
-      cryptoGiving,
-      setOffer,
-      offer,
-      cause,
-      setCause,
-      nonProfit,
-      setNonProfit,
-      flow,
-      setFlow,
-      loading,
-      resetStates,
     }),
-    [
-      currentCoin,
-      offer,
-      country,
-      city,
-      state,
-      taxId,
-      email,
-      number,
-      name,
-      expirationDate,
-      cvv,
-      buttonDisabled,
-      cause,
-      nonProfit,
-      flow,
-      loading,
-    ],
+    [handleSubmit],
   );
 
   return (
