+++ src/contexts/pixInformationContext/index.tsx
@@ -1,52 +1,26 @@
 import { useCurrentUser } from "contexts/currentUserContext";
 
-import {
-  createContext,
-  useContext,
-  SetStateAction,
-  useState,
-  useMemo,
-} from "react";
+import { createContext, useContext, useState, useMemo } from "react";
 import { useTranslation } from "react-i18next";
 import { logError } from "services/crashReport";
 import { useIntegration, useSources, useUsers } from "@ribon.io/shared/hooks";
 import pixPaymentApi from "services/api/pixPaymentApi";
 import PaymentIntent from "types/entities/PaymentIntent";
-import { ConfirmPaymentResult, useStripe } from "@stripe/stripe-react-native";
+import { ConfirmPaymentResult } from "@stripe/stripe-react-native";
 import { normalizedLanguage } from "lib/currentLanguage";
 import { PLATFORM } from "utils/constants/Application";
 import { showToast } from "lib/Toast";
 import { useNavigation } from "hooks/useNavigation";
 import { useIntegrationContext } from "contexts/integrationContext";
-import axios from "axios";
-import { Cause, NonProfit, Offer } from "@ribon.io/shared";
-import { countryByLanguage } from "lib/countryByLanguage";
-import { useLanguage } from "contexts/languageContext";
 import { useLoadingOverlay } from "contexts/loadingOverlayContext";
+import { useCheckoutContext } from "contexts/checkoutContext";
 
 export interface IPixPaymentInformationContext {
-  setButtonDisabled: (value: SetStateAction<boolean>) => void;
   buttonDisabled: boolean;
   handleSubmit: () => void;
   clientSecret?: string;
   pixInstructions?: PaymentIntent & ConfirmPaymentResult;
   verifyPayment: (clientSecret?: string, interval?: string) => void;
-  setCountry: (value: SetStateAction<string>) => void;
-  setTaxId: (value: SetStateAction<string>) => void;
-  setEmail: (value: SetStateAction<string | undefined>) => void;
-  setName: (value: SetStateAction<string>) => void;
-  setOffer: (value: SetStateAction<Offer | undefined>) => void;
-  setFlow: (value: SetStateAction<"cause" | "nonProfit">) => void;
-  country: string;
-  taxId: string;
-  email: string | undefined;
-  name: string;
-  offer: Offer | undefined;
-  flow: "cause" | "nonProfit";
-  cause: Cause | undefined;
-  setCause: (value: SetStateAction<Cause | undefined>) => void;
-  nonProfit: NonProfit | undefined;
-  setNonProfit: (value: SetStateAction<NonProfit | undefined>) => void;
 }
 
 export type Props = {
@@ -59,29 +33,12 @@ export const PixPaymentInformationContext =
   );
 
 function PixPaymentInformationProvider({ children }: Props) {
-  const { currentLang } = useLanguage();
-  const currentContry = countryByLanguage(currentLang);
-  const { currentUser } = useCurrentUser();
-
   const { showLoadingOverlay, hideLoadingOverlay } = useLoadingOverlay();
-  const [clientSecret, setClientSecret] = useState();
-  const [country, setCountry] = useState<string>(currentContry);
-  const [taxId, setTaxId] = useState<string>("");
-  const [email, setEmail] = useState<string | undefined>(
-    currentUser?.email ?? undefined,
-  );
-  const [name, setName] = useState<string>("");
-  const [buttonDisabled, setButtonDisabled] = useState(true);
-  const [offer, setOffer] = useState<Offer | undefined>();
-  const [cause, setCause] = useState<Cause | undefined>();
-  const [nonProfit, setNonProfit] = useState<NonProfit | undefined>();
-  const [flow, setFlow] = useState<"cause" | "nonProfit">("nonProfit");
 
   const { t } = useTranslation("translation", {
     keyPrefix: "contexts.pixPaymentInformation",
   });
 
-  const { confirmPayment } = useStripe();
   const { findOrCreateUser } = useUsers();
   const { signedIn, setCurrentUser } = useCurrentUser();
   const { currentIntegrationId } = useIntegrationContext();
@@ -93,12 +50,19 @@ function PixPaymentInformationProvider({ children }: Props) {
 
   const { navigateTo } = useNavigation();
 
-  const stripeApiKey =
-    "sk_test_51JRgaRJuOnwQq9Qxy7krZHxhvBXkNnydWiXQ89CNg94TxASNpOondjVSf6Qh2ooM8lA35KsIjy1aX2KGQas9fgJd00vNPkFAx0";
-  const headers = {
-    Authorization: `Bearer ${stripeApiKey}`,
-    "Content-Type": "application/x-www-form-urlencoded",
-  };
+  const [clientSecret, setClientSecret] = useState();
+  const {
+    email,
+    name,
+    buttonDisabled,
+    setButtonDisabled,
+    cause,
+    nonProfit,
+    offer,
+    taxId,
+    flow,
+    country,
+  } = useCheckoutContext();
 
   const login = async () => {
     if (!signedIn) {
@@ -116,13 +80,9 @@ function PixPaymentInformationProvider({ children }: Props) {
       }
     }
   };
-  const generatePixPayment = async (clientS: string) => {
+  const generatePixPayment = async (secret: string) => {
     try {
-      const response = await axios.post(
-        `https://api.stripe.com/v1/payment_intents/${clientS}/confirm`,
-        {},
-        { headers },
-      );
+      const response = await pixPaymentApi.generatePixPayment(secret);
 
       setPixInstructions(response.data);
       setTimeout(() => {
@@ -137,55 +97,56 @@ function PixPaymentInformationProvider({ children }: Props) {
     } catch (e) {
       logError(e);
 
+      hideLoadingOverlay();
+
       showToast({
         message: t("onErrorMessage"),
         type: "info",
       });
     } finally {
       setButtonDisabled(false);
+      hideLoadingOverlay();
     }
   };
 
-  const verifyPixPayment = async (secret?: string, interval?: string) => {
+  const verifyPixPayment = async (id?: string, interval?: string) => {
     try {
-      const { paymentIntent } = await confirmPayment(secret ?? "");
-
-      if (paymentIntent?.status === "Succeeded") {
-        clearInterval(interval);
-        login();
-        navigateTo("DonationDoneScreen", {
-          hasButton: true,
-          offerId: offer?.id ?? 0,
-          cause,
-          nonProfit,
-          flow,
-        });
-      }
+      await pixPaymentApi.verifyPixPayment(id ?? "").then((response) => {
+        if (response?.data.status === "succeeded") {
+          clearInterval(interval);
+          login();
+          navigateTo("ContributionDoneScreen", {
+            hasButton: true,
+            offerId: offer?.id ?? 0,
+            cause,
+            nonProfit,
+            flow,
+          });
+        }
+      });
     } catch (e) {
       logError(e);
-      showToast({
-        message: t("onErrorMessage"),
-        type: "info",
-      });
+
+      clearInterval(interval);
     }
   };
 
   const handleSubmit = async () => {
-    const paymentInformation = {
-      email: email ?? "",
-      country,
-      taxId: taxId ?? "",
-      offerId: offer?.id ?? 0,
-      name: name ?? "",
-      integrationId: currentIntegrationId ?? 1,
-      causeId: cause?.id,
-      nonProfitId: nonProfit?.id,
-      platform: PLATFORM,
-      city: "",
-      state: "",
-    };
-
     try {
+      const paymentInformation = {
+        email: email ?? "",
+        country,
+        taxId,
+        offerId: offer?.id ?? 0,
+        name,
+        integrationId: currentIntegrationId ?? 1,
+        causeId: cause?.id,
+        nonProfitId: nonProfit?.id,
+        platform: PLATFORM,
+        city: "",
+        state: "",
+      };
+
       const response = await pixPaymentApi.postPixPayment(paymentInformation);
 
       setClientSecret(response.data.externalId);
@@ -194,6 +155,7 @@ function PixPaymentInformationProvider({ children }: Props) {
     } catch (error) {
       hideLoadingOverlay();
       logError(error);
+
       showToast({
         message: t("onErrorMessage"),
         type: "info",
@@ -203,29 +165,12 @@ function PixPaymentInformationProvider({ children }: Props) {
 
   const pixPaymentInformationObject: IPixPaymentInformationContext = useMemo(
     () => ({
-      country,
-      taxId,
-      email,
-      name,
-      offer,
-      flow,
-      cause,
-      nonProfit,
-      setCause,
-      setNonProfit,
-      setButtonDisabled,
-      setCountry,
-      setTaxId,
-      setEmail,
-      setName,
-      setOffer,
-      setFlow,
       handleSubmit,
       buttonDisabled,
       pixInstructions,
       verifyPayment: verifyPixPayment,
     }),
-    [buttonDisabled, clientSecret],
+    [buttonDisabled, clientSecret, handleSubmit, pixInstructions],
   );
 
   return (
